feat(collections): create a collection to store and handle movie list functions
Implement a collection that store YoutubeMovie and execute function on it. The
functions implemented was:
- common_words that process all movie title and description and return the most
common words, to do this was created a string_sanitize function that remove
special characters from strings. Python Counter collection was used to count
word occurences.
- days_to_watch_all that count how many days is required to watch all movies in
the list based on weekday availability.Itertools cycle was used to make week
availability cyclic.